Refactor imports to use updated helpers structure
