+++ functional/metadata.test.ts
@@ -1,5 +1,5 @@
 import { closeEnv, loadEnv } from "@helpers/client";
-import { verifyStream } from "@helpers/tests";
+import { verifyStream } from "@helpers/streams";
 import { getWorkers, type WorkerManager } from "@workers/manager";
 import type { Group } from "@xmtp/node-sdk";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
