+++ functional/order.test.ts
@@ -1,10 +1,7 @@
 import { closeEnv, loadEnv } from "@helpers/client";
-import { getWorkersFromGroup } from "@helpers/group";
-import {
-  calculateMessageStats,
-  verifyStream,
-  type VerifyStreamResult,
-} from "@helpers/tests";
+import { getWorkersFromGroup } from "@helpers/groups";
+import { verifyStream, type VerifyStreamResult } from "@helpers/streams";
+import { calculateMessageStats } from "@helpers/tests";
 import { getWorkers, type WorkerManager } from "@workers/manager";
 import type { Group } from "@xmtp/node-sdk";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
